Q=40961 Had to use long for this one

+++ NTT.c
@@ -8,23 +8,25 @@
 #include "params.h"
 #include "multiplication.h"
 
-void printarray(int *array, int n)
+
+
+void printarray(long *array, long n)
 {
     printf("{");
-    for(int i=0; i<n;i++){
+    for(long i=0; i<n;i++){
         printf("%d, ",array[i]);
     }
     printf("}\n");
 }
 
-void initiate_NTT_forward(int *roots, int *NTT_forward, int n, int move, bool direction , int start, int  stop){
+void initiate_NTT_forward(long *roots, long *NTT_forward, long n, long move, bool direction , long start, long stop){
     if (start==stop){
         return;
     }
     else{
         NTT_forward[0] = roots[n/2];
-        int current = start +1;
-        int next_move= move*2;
+        long current = start +1;
+        long next_move= move*2;
         if(direction){
             initiate_NTT_forward(roots, NTT_forward + next_move, n / 2, next_move, true, current, stop);
             initiate_NTT_forward(roots, NTT_forward + next_move + 1, n / 2 + PRIMITIVE_N / 2, next_move + 1, false,
@@ -37,13 +39,13 @@ void initiate_NTT_forward(int *roots, int *NTT_forward, int n, int move, bool di
     }
 }
 
-void initiate_NTT_roots(int *NTT_forward, int level, int *NTT_roots){
-    int polynomials=1;
-    for (int i=0;i<level-1; i++){
+void initiate_NTT_roots(long *NTT_forward, long level, long *NTT_roots){
+    long polynomials=1;
+    for (long i=0;i<level-1; i++){
         polynomials=polynomials*2;
     }
     //printf("%d", polynomials);
-    for (int i =0; i<polynomials;i++){
+    for (long i =0; i<polynomials;i++){
         NTT_roots[i*2]=-NTT_forward[i+polynomials-1];
         NTT_roots[i*2+1]=NTT_forward[i+polynomials-1];
     }
@@ -59,10 +61,10 @@ void initiate_NTT_roots(int *NTT_forward, int level, int *NTT_roots){
  *
  * @author Anna Bakkebø
  */
-void splitting(int *pol, int n, int w){
+void splitting(long *pol, long n, long w){
 
-    int a;
-    for(int i=0; i<n; i++){
+    long a;
+    for(long i=0; i<n; i++){
         a = pol[i+n]*w;//printf("%d, ",pol[i]);
         pol[n+i]= (pol[i]-a)%Q;
         pol[i]=(pol[i]+a)%Q;
@@ -83,9 +85,9 @@ void splitting(int *pol, int n, int w){
  *
  * @author Anna Bakkebø
  */
-void merging(int * pol, int n, int w){
-    int a;
-    for (int i =0; i < n; i++){
+void merging(long * pol, long n, long w){
+    long a;
+    for (long i =0; i < n; i++){
         a = pol[i]+pol[n+i];
         pol[n+i]= ((pol[n+i]-pol[i])*w)%Q;
         pol[i]=a;
@@ -97,7 +99,7 @@ void merging(int * pol, int n, int w){
 
 }
 
-void forward_NTT(int *pol,int  i,int  n,int  m, int *roots, int start, int stop){
+void forward_NTT(long *pol, long i,long  n,long m, long * roots, long start, long stop){
     if(start==stop){
         return;
     }
@@ -109,14 +111,14 @@ void forward_NTT(int *pol,int  i,int  n,int  m, int *roots, int start, int stop)
         }*/
         if(start==stop){
             printf("Performing the last step of the forward NTT splitting {");
-            for(int i=0;i<n;i++){
+            for(long i=0;i<n;i++){
                 printarray(pol,n);
             }printf("using %d\n",roots[i]);
         }
         splitting(pol, n, roots[i]);
         //printf("performing splitting mod %d\n",roots[i]);
     }
-    int current=start+1;
+    long current=start+1;
     i=i/2;
     n=n/2;
     forward_NTT(pol, i,n,m,roots, current, stop);
@@ -125,7 +127,7 @@ void forward_NTT(int *pol,int  i,int  n,int  m, int *roots, int start, int stop)
 
 
 
-void forward_NTT2(int *pol, int *NTT_forward,int move, int start, int levels, int n){
+void forward_NTT2(long *pol, long *NTT_forward,long move, long start, long levels, long n){
     if(start==levels){
         return;
     }
@@ -140,7 +142,7 @@ void forward_NTT2(int *pol, int *NTT_forward,int move, int start, int levels, in
     }
     else{
         move=move*2;
-        for(int i=0; i<move;i++){
+        for(long i=0; i<move;i++){
             /*if(start==levels){
                 printf("Performing the last step of the forward NTT splitting {");
                 printarray(pol+i*n,n);
@@ -161,14 +163,14 @@ void forward_NTT2(int *pol, int *NTT_forward,int move, int start, int levels, in
 
 }
 
-void innverse_NTT2(int *pol, int *NTT_forward,int move, int start, int levels, int n){
+void innverse_NTT2(long *pol, long *NTT_forward,long move, long start, long levels, long n){
     if(start==levels){
         return;
     }
 
-    int i=move-1;
+    long i=move-1;
     //printf("Working on the %dth level:\n",start);
-    for(int j=0;j<move;j++){
+    for(long j=0;j<move;j++){
         //printf("Merging ");
         //printarray(pol+j*n,n);
         //printf(" using %d\n",NTT_forward[i]);
@@ -186,8 +188,8 @@ void innverse_NTT2(int *pol, int *NTT_forward,int move, int start, int levels, i
     innverse_NTT2(pol, NTT_forward-move,move, start, levels,  n*2);
 }
 
-void innverse_finnish(int *pol){
-    for(int i=0; i<N; i++){
+void innverse_finnish(long *pol){
+    for(long i=0; i<N; i++){
         pol[i]=(pol[i]*INNVERSE_POWER_OF_TWO)%Q;
 #if COUNTOPERATIONS==1
         Mult_NTT+=1;
@@ -197,14 +199,14 @@ void innverse_finnish(int *pol){
 
 void printNTTRoots(){
     printf("The NTT roots are {");
-    for (int i =0;i<NUM_POLYNOMIALS;i++){
+    for (long i =0;i<NUM_POLYNOMIALS;i++){
         printf("%d, " ,NTT_roots[i]);
     }
     printf("}\n");
 }
 void printNTT_Forward(){
     printf("The NTT forward are {");
-    for (int i =0;i<NUM_POLYNOMIALS;i++){
+    for (long i =0;i<NUM_POLYNOMIALS;i++){
         printf("%d, " ,NTT_forward[i]);
     }
     printf("}\n");
